Add macronutrients to report and change calorie intake type to int

+++ types.ml
@@ -52,6 +52,6 @@ type user =
   height : int;
   goal : goal;
   pal : float;
-  calorie_intake : float;
+  calorie_intake : int;
   days : day list
 }
